curso-4-3-2-langchain-crie-agentes-ia-apps-com-llms-e-llamaindex/pyproject.toml
[project]
name = "curso-4-3-2-langchain-crie-agentes-ia-apps-com-llms-e-llamaindex"
version = "0.1.0"
description = ""
authors = [
    {name = "Sóstenes Souza",email = "sostenes198br@hotmail.com"}
]
requires-python = "=3.12.10"
dependencies = [
    "groq (>=1.1.2,<2.0.0)",
    "python-dotenv (>=1.2.2,<2.0.0)",
    "chromadb (>=1.5.7,<2.0.0)",
    "sentence-transformers (>=5.4.0,<6.0.0)",
    "langchain (>=1.2.15,<2.0.0)",
    "langchain-openai (>=1.1.13,<2.0.0)",
    "ipykernel (>=7.2.0,<8.0.0)",
    "langchain-community (>=0.4.1,<0.5.0)",
    "openai (>=2.32.0,<3.0.0)",
    "huggingface-hub (>=1.10.2,<2.0.0)",
    "transformers (>=5.5.4,<6.0.0)",
    "jinja2 (>=3.1.6,<4.0.0)",
    "tiktoken (>=0.12.0,<0.13.0)",
    "pypdf (>=6.10.2,<7.0.0)",
    "yt-dlp (>=2026.3.17,<2027.0.0)",
    "pydub (>=0.25.1,<0.26.0)",
    "beautifulsoup4 (>=4.14.3,<5.0.0)",
    "langchain-chroma (>=1.1.0,<2.0.0)",
    "faiss-cpu (>=1.13.2,<2.0.0)",
    "static-ffmpeg (>=3.0,<4.0)",
    "ipywidgets (>=8.1.8,<9.0.0)",
    "playwright (>=1.58.0,<2.0.0)",
    "unstructured (>=0.22.22,<0.23.0)"
]

[tool.poetry]
package-mode = false


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "jupyterlab (>=4.5.6,<5.0.0)"
]
